:ok_hand: Updating code due to code review changes.

+++ src/main/java/in/gaks/oneyard/model/entity/Material.java
@@ -1,9 +1,13 @@
 package in.gaks.oneyard.model.entity;
 
 import in.gaks.oneyard.base.BaseEntity;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 import lombok.experimental.Accessors;
-import lombok.*;
-import javax.persistence.*;
 import org.hibernate.annotations.Where;
 
 
@@ -30,12 +34,12 @@ public class Material extends BaseEntity {
   /**
    * 物料类别id.
    */
-  private java.lang.Integer typeId;
+  private java.lang.Long typeId;
 
   /**
    * 物料库存数量.
    */
-  private java.lang.Integer number;
+  private java.lang.Long number;
 
   /**
    * 规格.
